+++ app/src/main/res/layout-sw600dp-land/layout_main_dz6.xml
@@ -1,0 +1,26 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="horizontal"
+    android:baselineAligned="false">
+
+    <fragment
+        android:id="@+id/fragmentPersonList"
+        android:name="com.golda.app.pvttraining.dz6.PersonsLiatFragment"
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        android:layout_weight="1"
+        tools:layout="@layout/fragment_peson_list"/>
+
+    <fragment
+        android:id="@+id/fragmentEditPerson"
+        android:name="com.golda.app.pvttraining.dz6.EditPersonFragment"
+        android:layout_width="0dp"
+        android:layout_height="match_parent"
+        android:layout_weight="2"
+        tools:layout="@layout/fragment_edit_person"/>
+
+</LinearLayout>
